Separated all actions related to a specific component into a new file
