+++ client/scripts/reducers/bankId.js
@@ -1,4 +1,6 @@
-import { LoginStages, INPUT_PERSONAL_NUMBER, SET_PERSONAL_NUMBER_VALIDITY, SET_BID_STAGE, SET_BID_STATUS, SET_BID_ORDER, RESET_STATE, INCREMENT_BID_TRIES, SET_USER_NAME } from '../actions/actions';
+import { INPUT_PERSONAL_NUMBER, SET_PERSONAL_NUMBER_VALIDITY, SET_BID_STAGE, 
+	SET_BID_STATUS, SET_BID_ORDER, INCREMENT_BID_TRIES, SET_USER_NAME } from '../actions/bankid';
+import { LoginStages, RESET_STATE } from '../actions/login';
 
 const initialState = {
 	bidStage: LoginStages.INPUT_PERSONAL_NUMBER,
